Move api out of pattern
