+++ crates/cli/src/analyze.rs
@@ -15,11 +15,11 @@ use indicatif::{MultiProgress, ProgressBar, ProgressDrawTarget, ProgressStyle};
 
 #[allow(unused_imports)]
 use marzano_core::pattern::built_in_functions::BuiltIns;
-use marzano_core::pattern::{
-    api::{AnalysisLog, DoneFile, MatchResult},
-    Problem,
-};
 use marzano_core::pattern_compiler::{src_to_problem_libs, CompilationResult};
+use marzano_core::{
+    api::{AnalysisLog, DoneFile, MatchResult},
+    problem::Problem,
+};
 use marzano_language::target_language::PatternLanguage;
 use marzano_util::cache::GritCache;
 use marzano_util::position::{FileRange, Position};
