+++ crates/cli_bin/tests/snapshots/apply__gives_helpful_error_for_file.snap.new
@@ -1,0 +1,6 @@
+---
+source: crates/cli_bin/tests/apply.rs
+assertion_line: 1098
+expression: stdout
+---
+
