+++ crates/core/src/lib.rs
@@ -1,4 +1,5 @@
 #![deny(clippy::wildcard_enum_match_arm)]
+pub mod api;
 pub mod binding;
 pub mod compact_api;
 pub mod context;
@@ -12,6 +13,7 @@ mod orphan;
 pub mod parse;
 pub mod pattern;
 pub mod pattern_compiler;
+pub mod problem;
 mod resolve;
 mod smart_insert;
 mod split_snippet;
