+++ crates/core/src/pattern/step.rs
@@ -3,11 +3,11 @@ use super::{
     patterns::{Matcher, Pattern},
     resolved_pattern::{File, ResolvedPattern},
     state::{FilePtr, State},
-    FileOwner, InputRanges, MatchRanges,
 };
 use crate::{
     context::Context,
     orphan::{get_orphaned_ranges, remove_orphaned_ranges},
+    problem::{FileOwner, InputRanges, MatchRanges},
     text_unparser::apply_effects,
 };
 use anyhow::{anyhow, bail, Result};
