+++ crates/lsp/src/actions.rs
@@ -1,5 +1,8 @@
-use anyhow::{Result, anyhow};
-use marzano_core::{fs::extract_ranges, pattern::api::{make_suppress_comment, MatchResult}};
+use anyhow::{anyhow, Result};
+use marzano_core::{
+    api::{make_suppress_comment, MatchResult},
+    fs::extract_ranges,
+};
 use marzano_language::target_language::{PatternLanguage, TargetLanguage};
 use tower_lsp::lsp_types::{
     self, CodeAction, CodeActionKind, CodeActionOrCommand, Command, DocumentChanges, OneOf,
@@ -74,14 +77,20 @@ fn make_suppress_action(
     range: &Range,
     local_name: &str,
 ) -> Result<CodeAction> {
-    let language = language_id_to_pattern_language(&document.language_id).unwrap_or(PatternLanguage::JavaScript);
-    let target_language = TargetLanguage::try_from(language).map_err(|_| anyhow!("Invalid language"))?;
+    let language = language_id_to_pattern_language(&document.language_id)
+        .unwrap_or(PatternLanguage::JavaScript);
+    let target_language =
+        TargetLanguage::try_from(language).map_err(|_| anyhow!("Invalid language"))?;
     let insert_text = make_suppress_comment(Some(local_name), &target_language);
     let whitespace = document
         .text
         .lines()
         .nth(range.start.line as usize)
-        .map(|line| line.chars().take_while(|c| c.is_whitespace()).collect::<String>())
+        .map(|line| {
+            line.chars()
+                .take_while(|c| c.is_whitespace())
+                .collect::<String>()
+        })
         .unwrap_or_default();
     let insert_text = format!("{}{}", whitespace, insert_text);
     let text_edit = TextEdit {
