+++ crates/lsp/src/apply.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use marzano_core::pattern::api::MatchResult;
+use marzano_core::api::MatchResult;
 use marzano_core::pattern_compiler::src_to_problem_libs;
 use marzano_gritmodule::patterns_directory::PatternsDirectory;
 use marzano_language::target_language::PatternLanguage;
