+++ crates/wasm-bindings/src/match_pattern.rs
@@ -1,8 +1,6 @@
 use marzano_core::{
-    pattern::{
-        api::{AnalysisLog, InputFile, MatchResult, PatternInfo},
-        built_in_functions::BuiltIns,
-    },
+    api::{AnalysisLog, InputFile, MatchResult, PatternInfo},
+    pattern::built_in_functions::BuiltIns,
     pattern_compiler::{src_to_problem_libs_for_language, CompilationResult},
     tree_sitter_serde::tree_sitter_node_to_json,
 };
